Update.
2002-06-04  Bruno Haible  <bruno@clisp.org>

	* sysdeps/cris/stackinfo.h: New file.

	* sysdeps/hppa/dl-machine.h (elf_machine_rela): Cast reloc_addr
	to an integer type before performing bit operations on it.

+++ sysdeps/hppa/dl-machine.h
@@ -531,7 +531,7 @@ elf_machine_rela (struct link_map *map, const Elf32_Rela *reloc,
 	return;
 #endif
       /* .eh_frame can have unaligned relocs.  */
-      if (reloc_addr & 3)
+      if ((unsigned long) reloc_addr & 3)
 	{
 	  char *rel_addr = (char *) reloc_addr;
 	  rel_addr[0] = value >> 24;
